+++ src/utils/account/addressIsValid.ts
@@ -2,16 +2,14 @@ import { Address } from '@multiversx/sdk-core';
 
 function canTransformToPublicKey(address: string) {
   try {
-    const checkAddress = new Address(address);
-    return Boolean(checkAddress.bech32());
+    return Address.isValid(address);
   } catch {
     return false;
   }
 }
 
 export function addressIsValid(destinationAddress: string) {
-  const isValidBach =
-    destinationAddress?.length === 62 && /^\w+$/.test(destinationAddress);
+  const isValidString = /^\w+$/.test(destinationAddress);
 
-  return isValidBach && canTransformToPublicKey(destinationAddress);
+  return isValidString && canTransformToPublicKey(destinationAddress);
 }